Create README.md
I'm a developer who loves to learn new things and solve problems. I have been developing apps and websites for over 3 years and I enjoy every aspect of it. I'm always looking for new challenges and opportunities to improve my skills and create value for others. I'm also interested in artificial intelligence, machine learning, and data science. I believe that technology can make the world a better place and I want to be a part of it.